Make XLAT a independent library
- Moved XLAT library to tools/lib/ from val/common/
- Added acs_xlat.cmake as a helper file to build XLAT library along ACS
- Added tools/configs/xlat/xlat_config.h which is a prerequisite for
  XLAT library
- Added CMake files for XLAT library to be able to  build XLAT
  independently outside ACS
- Modified project CMake to build and link independently built  XLAT
  library
