+++ CMakeLists.txt
@@ -5,7 +5,7 @@
 #
 #-------------------------------------------------------------------------------
 
-project(rmm-acs LANGUAGES)
+project(rmm-acs LANGUAGES C ASM)
 
 ### Tool dependency check - start ###
 
@@ -249,6 +249,7 @@ endforeach()
 ###
 
 ### Compile host, realm and secure val/pal/test sources and create output binaries
+include(${ROOT_DIR}/tools/cmake/common/acs_xlat.cmake)
 add_subdirectory(${ROOT_DIR}/tools/cmake/acs_host)
 add_subdirectory(${ROOT_DIR}/tools/cmake/acs_realm)
 add_subdirectory(${ROOT_DIR}/tools/cmake/acs_secure)
