+++ test/database/test_secure.cmake
@@ -1,5 +1,5 @@
 #-------------------------------------------------------------------------------
-# Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+# Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
 #
 # SPDX-License-Identifier: BSD-3-Clause
 #
@@ -28,7 +28,6 @@ list(APPEND TEST_INCLUDE
     ${ROOT_DIR}/plat/driver/src/
     ${ROOT_DIR}/test/database/
     ${ROOT_DIR}/test/common/
-    ${ROOT_DIR}/val/common/xlat_tables_v2/include/
 )
 
 foreach(SUITE ${TEST_SUITE_LIST})
@@ -68,6 +67,8 @@ set(COMPILE_LIST ${COMPILE_LIST} PARENT_SCOPE)
 
 # Create TEST library
 add_library(${TEST_LIB} STATIC ${TEST_SRC})
+
+target_link_libraries(${TEST_LIB} PUBLIC ${XLAT-LIB})
 
 target_include_directories(${TEST_LIB} PRIVATE ${TEST_INCLUDE})
 
@@ -111,6 +112,7 @@ foreach(SUITE ${SUITE_LIST})
         add_library(${TEST_LIB} STATIC ${TEST_SRC})
         target_compile_definitions(${TEST_LIB} PRIVATE d_${TEST})
         target_include_directories(${TEST_LIB} PRIVATE ${TEST_INCLUDE})
+        target_link_libraries(${TEST_LIB} PUBLIC ${XLAT-LIB})
         file(MAKE_DIRECTORY ${BUILD}/output/${SUITE}/${TEST})
         create_executable(${EXE_NAME} ${BUILD}/output/${SUITE}/${TEST} ${TEST})
         remove_definitions(-Dd_${TEST})
