+++ tools/configs/xlat/xlat_config.h
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ */
+
+#define PLAT_ASSERT_FUNC val_assert
+#define PLAT_PRINT_FUNC val_common_printf
