+++ tools/lib/xlat_tables_v2/CMakeLists.txt
@@ -1,0 +1,17 @@
+#-------------------------------------------------------------------------------
+# Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+#
+#-------------------------------------------------------------------------------
+
+
+file(GLOB XLAT_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/*.S"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/*.c"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/*.S"
+)
+
+add_library(${XLAT-LIB} STATIC ${XLAT_SRC})
+
+target_include_directories(${XLAT-LIB} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include/)
