+++ tools/lib/xlat_tables_v2/include/arch_features.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -11,37 +11,36 @@
 #define ARCH_FEATURES_H
 
 #include <stdbool.h>
-#include <val_sysreg.h>
-#include <pal_arch.h>
-#include <val.h>
+#include <arch.h>
+#include <arch_helpers.h>
 
 static inline unsigned int read_id_aa64mmfr0_el0_tgran64_field(void)
 {
-	return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN64,
-		val_id_aa64mmfr0_el1_read()));
+    return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN64,
+        read_id_aa64mmfr0_el1()));
 }
 static inline unsigned int read_id_aa64mmfr0_el0_tgran16_field(void)
 {
-	return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN16,
-		val_id_aa64mmfr0_el1_read()));
+    return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN16,
+        read_id_aa64mmfr0_el1()));
 }
 
 static inline unsigned int read_id_aa64mmfr0_el0_tgran4_field(void)
 {
-	return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN4,
-		val_id_aa64mmfr0_el1_read()));
+    return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN4,
+        read_id_aa64mmfr0_el1()));
 }
 
 static inline bool is_armv8_4_ttst_present(void)
 {
-	return (EXTRACT(ID_AA64MMFR2_EL1_ST,
-		val_id_aa64mmfr2_el1_read()) == 1U);
+    return (EXTRACT(ID_AA64MMFR2_EL1_ST,
+        read_id_aa64mmfr2_el1()) == 1U);
 }
 
 static inline bool is_armv8_2_ttcnp_present(void)
 {
-	return (EXTRACT(ID_AA64MMFR2_EL1_CNP,
-		val_id_aa64mmfr2_el1_read()) == 1U);
+    return (EXTRACT(ID_AA64MMFR2_EL1_CNP,
+        read_id_aa64mmfr2_el1()) == 1U);
 }
 
 /*
@@ -52,8 +51,8 @@ static inline bool is_armv8_2_ttcnp_present(void)
  */
 static inline bool is_feat_sve_present(void)
 {
-	return (EXTRACT(ID_AA64PFR0_EL1_SVE,
-		val_id_aa64pfr0_el1_read()) != 0UL);
+    return (EXTRACT(ID_AA64PFR0_EL1_SVE,
+        read_id_aa64pfr0_el1()) != 0UL);
 }
 
 /*
@@ -61,8 +60,8 @@ static inline bool is_feat_sve_present(void)
  */
 static inline bool is_feat_rng_present(void)
 {
-	return (EXTRACT(ID_AA64ISAR0_EL1_RNDR,
-		val_id_aa64isar0_el1_read()) != 0UL);
+    return (EXTRACT(ID_AA64ISAR0_EL1_RNDR,
+        read_id_aa64isar0_el1()) != 0UL);
 }
 
 /*
@@ -74,8 +73,8 @@ static inline bool is_feat_rng_present(void)
  */
 static inline bool is_feat_vmid16_present(void)
 {
-	return (EXTRACT(ID_AA64MMFR1_EL1_VMIDBits,
-		val_id_aa64mmfr1_el1_read()) == ID_AA64MMFR1_EL1_VMIDBits_16);
+    return (EXTRACT(ID_AA64MMFR1_EL1_VMIDBits,
+        read_id_aa64mmfr1_el1()) == ID_AA64MMFR1_EL1_VMIDBits_16);
 }
 
 /*
@@ -85,8 +84,8 @@ static inline bool is_feat_vmid16_present(void)
  */
 static inline bool is_feat_lpa2_4k_present(void)
 {
-	return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN4,
-		val_id_aa64mmfr0_el1_read()) == ID_AA64MMFR0_EL1_TGRAN4_LPA2);
+    return (EXTRACT(ID_AA64MMFR0_EL1_TGRAN4,
+        read_id_aa64mmfr0_el1()) == ID_AA64MMFR0_EL1_TGRAN4_LPA2);
 }
 
 /*
@@ -98,22 +97,12 @@ static inline bool is_feat_lpa2_4k_present(void)
  */
 static inline bool is_feat_lpa2_4k_2_present(void)
 {
-	u_register_t id_aa64mmfr0_el1 = val_id_aa64mmfr0_el1_read();
+    u_register_t id_aa64mmfr0_el1 = read_id_aa64mmfr0_el1();
 
-	return ((EXTRACT(ID_AA64MMFR0_EL1_TGRAN4_2, id_aa64mmfr0_el1) ==
-		ID_AA64MMFR0_EL1_TGRAN4_2_LPA2) ||
-		 ((EXTRACT(ID_AA64MMFR0_EL1_TGRAN4_2, id_aa64mmfr0_el1) ==
-		ID_AA64MMFR0_EL1_TGRAN4_2_TGRAN4) && is_feat_lpa2_4k_present()));
-}
-
-/*
- * Returns Performance Monitors Extension version.
- * ID_AA64DFR0_EL1.PMUVer, bits [11:8]:
- * 0b0000: Performance Monitors Extension not implemented
- */
-static inline unsigned int read_pmu_version(void)
-{
-	return EXTRACT(ID_AA64DFR0_EL1_PMUVer, val_id_aa64dfr0_el1_read());
+    return ((EXTRACT(ID_AA64MMFR0_EL1_TGRAN4_2, id_aa64mmfr0_el1) ==
+        ID_AA64MMFR0_EL1_TGRAN4_2_LPA2) ||
+         ((EXTRACT(ID_AA64MMFR0_EL1_TGRAN4_2, id_aa64mmfr0_el1) ==
+        ID_AA64MMFR0_EL1_TGRAN4_2_TGRAN4) && is_feat_lpa2_4k_present()));
 }
 
 unsigned int arch_feat_get_pa_width(void);
