+++ tools/lib/xlat_tables_v2/include/config.h
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ */
+
+#if defined(XLAT_CONFIG_FILE)
+#include XLAT_CONFIG_FILE
+#else
+#error "Please include -DXLAT_CONFIG_FILE during compilation"
+#endif
