+++ tools/lib/xlat_tables_v2/include/xlat_mmu_helpers.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -21,7 +21,7 @@
  * for all levels of data and unified cache until the caches are enabled by
  * setting the bit SCTLR_ELx.C.
  */
-#define DISABLE_DCACHE			(U(1) << 0)
+#define DISABLE_DCACHE            (U(1) << 0)
 
 /*
  * Mark the translation tables as non-cacheable for the MMU table walker, which
@@ -42,16 +42,16 @@
  * that there aren't mismatches. The behaviour in that case is described in the
  * sections 'Mismatched memory attributes' in the ARMv8 ARM.
  */
-#define XLAT_TABLE_NC			(U(1) << 1)
+#define XLAT_TABLE_NC            (U(1) << 1)
 
 /*
  * Offsets into a mmu_cfg_params array generated by setup_mmu_cfg(). All
  * parameters are 64 bits wide.
  */
-#define MMU_CFG_MAIR		0
-#define MMU_CFG_TCR		1
-#define MMU_CFG_TTBR0		2
-#define MMU_CFG_PARAM_MAX	3
+#define MMU_CFG_MAIR        0
+#define MMU_CFG_TCR        1
+#define MMU_CFG_TTBR0        2
+#define MMU_CFG_PARAM_MAX    3
 
 #ifndef __ASSEMBLER__
 
@@ -66,8 +66,8 @@
  * MMU_CFG_PARAM_MAX.
  */
 void setup_mmu_cfg(uint64_t *params, unsigned int flags,
-		   const uint64_t *base_table, unsigned long long max_pa,
-		   uintptr_t max_va, int xlat_regime);
+           const uint64_t *base_table, unsigned long long max_pa,
+           uintptr_t max_va, int xlat_regime);
 
 #ifdef __aarch64__
 /* AArch64 specific translation table APIs */
