+++ tools/lib/xlat_tables_v2/include/xlat_tables.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -31,14 +31,14 @@
 /*
  * Shifts and masks to access fields of an mmap attribute
  */
-#define MT_TYPE_MASK	U(0x7)
-#define MT_TYPE(_attr)	((_attr) & MT_TYPE_MASK)
+#define MT_TYPE_MASK    U(0x7)
+#define MT_TYPE(_attr)    ((_attr) & MT_TYPE_MASK)
 /* Access permissions (RO/RW) */
-#define MT_PERM_SHIFT	U(3)
+#define MT_PERM_SHIFT    U(3)
 /* Security state (SECURE/NS) */
-#define MT_SEC_SHIFT	U(4)
+#define MT_SEC_SHIFT    U(4)
 /* Access permissions for instruction execution (EXECUTE/EXECUTE_NEVER) */
-#define MT_EXECUTE_SHIFT	U(5)
+#define MT_EXECUTE_SHIFT    U(5)
 
 /*
  * Memory mapping attributes
@@ -50,16 +50,16 @@
  * getting weaker; conversely going up the list the memory types are getting
  * stronger.
  */
-#define MT_DEVICE		U(0)
-#define MT_NON_CACHEABLE	U(1)
-#define MT_MEMORY		U(2)
+#define MT_DEVICE        U(0)
+#define MT_NON_CACHEABLE    U(1)
+#define MT_MEMORY        U(2)
 /* Values up to 7 are reserved to add new memory types in the future */
 
-#define MT_RO			(U(0) << MT_PERM_SHIFT)
-#define MT_RW			(U(1) << MT_PERM_SHIFT)
+#define MT_RO            (U(0) << MT_PERM_SHIFT)
+#define MT_RW            (U(1) << MT_PERM_SHIFT)
 
-#define MT_SECURE		(U(0) << MT_SEC_SHIFT)
-#define MT_NS			(U(1) << MT_SEC_SHIFT)
+#define MT_SECURE        (U(0) << MT_SEC_SHIFT)
+#define MT_NS            (U(1) << MT_SEC_SHIFT)
 
 /*
  * Access permissions for instruction execution are only relevant for normal
@@ -68,34 +68,34 @@
  *  - Device memory is always marked as execute-never.
  *  - Read-write normal memory is always marked as execute-never.
  */
-#define MT_EXECUTE		(U(0) << MT_EXECUTE_SHIFT)
-#define MT_EXECUTE_NEVER	(U(1) << MT_EXECUTE_SHIFT)
+#define MT_EXECUTE        (U(0) << MT_EXECUTE_SHIFT)
+#define MT_EXECUTE_NEVER    (U(1) << MT_EXECUTE_SHIFT)
 
 /* Compound attributes for most common usages */
-#define MT_CODE			(MT_MEMORY | MT_RO | MT_EXECUTE)
-#define MT_RO_DATA		(MT_MEMORY | MT_RO | MT_EXECUTE_NEVER)
+#define MT_CODE            (MT_MEMORY | MT_RO | MT_EXECUTE)
+#define MT_RO_DATA        (MT_MEMORY | MT_RO | MT_EXECUTE_NEVER)
 
 /* Memory type for EL3 regions */
 #if ENABLE_RME
 #error FEAT_RME requires version 2 of the Translation Tables Library
 #else
-#define EL3_PAS			MT_SECURE
+#define EL3_PAS            MT_SECURE
 #endif
 
 /*
  * Structure for specifying a single region of memory.
  */
 typedef struct mmap_region {
-	unsigned long long	base_pa;
-	uintptr_t		base_va;
-	size_t			size;
-	unsigned int		attr;
+    unsigned long long    base_pa;
+    uintptr_t        base_va;
+    size_t            size;
+    unsigned int        attr;
 } mmap_region_t;
 
 /* Generic translation table APIs */
 void init_xlat_tables(void);
 void mmap_add_region(unsigned long long base_pa, uintptr_t base_va,
-		     size_t size, unsigned int attr);
+             size_t size, unsigned int attr);
 void mmap_add(const mmap_region_t *mm);
 
 #endif /*__ASSEMBLER__*/
