+++ tools/lib/xlat_tables_v2/include/xlat_tables_arch.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -20,15 +20,15 @@
  * Evaluates to 1 if the given physical address space size is a power of 2,
  * or 0 if it's not.
  */
-#define CHECK_PHY_ADDR_SPACE_SIZE(size)				\
-	(IS_POWER_OF_TWO(size))
+#define CHECK_PHY_ADDR_SPACE_SIZE(size)                \
+    (IS_POWER_OF_TWO(size))
 
 /*
  * Compute the number of entries required at the initial lookup level to address
  * the whole virtual address space.
  */
-#define GET_NUM_BASE_LEVEL_ENTRIES(addr_space_size)			\
-	((addr_space_size) >>						\
-		XLAT_ADDR_SHIFT(GET_XLAT_TABLE_LEVEL_BASE(addr_space_size)))
+#define GET_NUM_BASE_LEVEL_ENTRIES(addr_space_size)            \
+    ((addr_space_size) >>                        \
+        XLAT_ADDR_SHIFT(GET_XLAT_TABLE_LEVEL_BASE(addr_space_size)))
 
 #endif /* XLAT_TABLES_ARCH_H */
